requirements.txt
tcex>=0.5,<0.6
dnspython